+++ src/modules/tools/wfsSearch/tests/unit/components/WfsSearch.spec.js
@@ -16,7 +16,12 @@ describe("src/modules/tools/wfsSearch/components/WfsSearch.vue", () => {
         },
         instanceChangedOrig = WfsSearchModule.actions.instanceChanged;
     let instances,
-        store;
+        store,
+        placingPointMarkerSpy,
+        placingPolygonMarkerSpy,
+        setCenterSpy,
+        setZoomLevelSpy,
+        zoomToExtentSpy;
 
     beforeEach(() => {
         const map = {
@@ -27,6 +32,12 @@ describe("src/modules/tools/wfsSearch/components/WfsSearch.vue", () => {
 
         mapCollection.clear();
         mapCollection.addMap(map, "2D");
+
+        placingPointMarkerSpy = sinon.spy();
+        placingPolygonMarkerSpy = sinon.spy();
+        setCenterSpy = sinon.spy();
+        setZoomLevelSpy = sinon.spy();
+        zoomToExtentSpy = sinon.spy();
 
         instances = [{
             title: "Test WfsSearch",
@@ -46,7 +57,17 @@ describe("src/modules/tools/wfsSearch/components/WfsSearch.vue", () => {
                 MapMarker: {
                     namespaced: true,
                     actions: {
-                        removePointMarker: sinon.stub()
+                        removePointMarker: sinon.stub(),
+                        placingPointMarker: placingPointMarkerSpy,
+                        placingPolygonMarker: placingPolygonMarkerSpy
+                    }
+                },
+                Maps: {
+                    namespaced: true,
+                    actions: {
+                        setCenter: setCenterSpy,
+                        setZoomLevel: setZoomLevelSpy,
+                        zoomToExtent: zoomToExtentSpy
                     }
                 },
                 Alerting: {
@@ -249,5 +270,82 @@ describe("src/modules/tools/wfsSearch/components/WfsSearch.vue", () => {
             pagination = wrapper.find("ul.pagination");
 
         expect(pagination.exists()).to.be.false;
+    });
+
+    describe("markerAndZoom", () => {
+        let pointFeatures,
+            polygonFeatures;
+
+        beforeEach(() => {
+            pointFeatures = [
+                {
+                    getGeometry: () => {
+                        return {
+                            getCoordinates: () => [568366.068, 5941065.428]
+                        };
+                    }
+                }
+            ];
+            polygonFeatures = [
+                {
+                    getGeometry: () => {
+                        return {
+                            getCoordinates: () => [
+                                [456881.4, 5341325.7, 0],
+                                [456905.5, 5341311.3, 0],
+                                [456931.2, 5341295.9, 0],
+                                [456932.3, 5341295.6, 0],
+                                [456936.2, 5341294.5, 0],
+                                [456940, 5341301.7, 0],
+                                [456943.2, 5341308, 0],
+                                [456946.6, 5341314.5, 0],
+                                [456949.4, 5341319.9, 0],
+                                [456929.5, 5341329.8, 0],
+                                [456914, 5341337.5, 0],
+                                [456893, 5341347.9, 0],
+                                [456882.2, 5341327.1, 0],
+                                [456881.4, 5341325.7, 0]
+                            ],
+                            getExtent: () => [456881.4, 5341294.5, 456949.4, 5341347.9]
+                        };
+                    }
+                }
+            ];
+        });
+        it("should start action placingPointMarker, if the feature has a point geometry", () => {
+            const wrapper = mount(WfsSearch, {
+                localVue,
+                store,
+                stubs: ["WfsSearchLiteral"]
+            });
+
+            wrapper.vm.markerAndZoom(pointFeatures);
+
+            expect(placingPointMarkerSpy.calledOnce).to.be.true;
+            expect(setCenterSpy.calledOnce).to.be.true;
+            expect(setZoomLevelSpy.calledOnce).to.be.true;
+
+            expect(placingPointMarkerSpy.firstCall.args[1]).to.deep.equals([568366.068, 5941065.428]);
+            expect(setCenterSpy.firstCall.args[1]).to.deep.equals([568366.068, 5941065.428]);
+            expect(setZoomLevelSpy.firstCall.args[1]).to.equals(5);
+        });
+
+        it("should start action placingPolygonMarker, if the feature has a polygon geometry", () => {
+            const wrapper = mount(WfsSearch, {
+                localVue,
+                store,
+                stubs: ["WfsSearchLiteral"]
+            });
+
+            wrapper.vm.markerAndZoom(polygonFeatures);
+
+            expect(placingPolygonMarkerSpy.calledOnce).to.be.true;
+            expect(zoomToExtentSpy.calledOnce).to.be.true;
+
+            expect(placingPolygonMarkerSpy.firstCall.args[1]).to.deep.equals(polygonFeatures[0]);
+            expect(zoomToExtentSpy.firstCall.args[1]).to.deep.equals({
+                extent: [456881.4, 5341294.5, 456949.4, 5341347.9]
+            });
+        });
     });
 });